requirements.txt
streamlit
pandas
scikit-learn
matplotlib
pdfplumber
pytesseract
pillow